Refactored: Legacy entities namespace.
