+++ app/config/doctrine/legacy/Config.orm.xml
@@ -1,6 +1,6 @@
 <?xml version="1.0" encoding="utf-8"?>
 <doctrine-mapping xmlns="http://doctrine-project.org/schemas/orm/doctrine-mapping" xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance" xsi:schemaLocation="http://doctrine-project.org/schemas/orm/doctrine-mapping http://doctrine-project.org/schemas/orm/doctrine-mapping.xsd">
-  <entity name="AppBundle\Entity\Config" table="config">
+  <entity name="AppBundle\Entity\Legacy\Config" table="config">
     <id name="key" type="string" column="key" length="255">
       <generator strategy="IDENTITY"/>
     </id>
