+++ app/config/doctrine/legacy/Runde.orm.xml
@@ -1,6 +1,6 @@
 <?xml version="1.0" encoding="utf-8"?>
 <doctrine-mapping xmlns="http://doctrine-project.org/schemas/orm/doctrine-mapping" xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance" xsi:schemaLocation="http://doctrine-project.org/schemas/orm/doctrine-mapping http://doctrine-project.org/schemas/orm/doctrine-mapping.xsd">
-  <entity name="AppBundle\Entity\Runde" table="runde">
+  <entity name="AppBundle\Entity\Legacy\Runde" table="runde">
     <indexes>
       <index name="runde_FI_1" columns="turnier_id"/>
     </indexes>
