+++ app/config/doctrine/legacy/Spiel.orm.xml
@@ -1,6 +1,6 @@
 <?xml version="1.0" encoding="utf-8"?>
 <doctrine-mapping xmlns="http://doctrine-project.org/schemas/orm/doctrine-mapping" xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance" xsi:schemaLocation="http://doctrine-project.org/schemas/orm/doctrine-mapping http://doctrine-project.org/schemas/orm/doctrine-mapping.xsd">
-  <entity name="AppBundle\Entity\Spiel" table="spiel">
+  <entity name="AppBundle\Entity\Legacy\Spiel" table="spiel">
     <indexes>
       <index name="spiel_FI_1" columns="turnier_id"/>
       <index name="spiel_FI_2" columns="runde_id"/>
