+++ app/config/doctrine/legacy/Tipp.orm.xml
@@ -1,6 +1,6 @@
 <?xml version="1.0" encoding="utf-8"?>
 <doctrine-mapping xmlns="http://doctrine-project.org/schemas/orm/doctrine-mapping" xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance" xsi:schemaLocation="http://doctrine-project.org/schemas/orm/doctrine-mapping http://doctrine-project.org/schemas/orm/doctrine-mapping.xsd">
-  <entity name="AppBundle\Entity\Tipp" table="tipp">
+  <entity name="AppBundle\Entity\Legacy\Tipp" table="tipp">
     <indexes>
       <index name="tipp_FI_1" columns="turnier_id"/>
       <index name="tipp_FI_2" columns="spieler_id"/>
