+++ src/AppBundle/Entity/Legacy/Config.php
@@ -1,6 +1,6 @@
 <?php
 
-namespace AppBundle\Entity;
+namespace AppBundle\Entity\Legacy;
 
 /**
  * Config
