+++ src/AppBundle/Entity/Legacy/Runde.php
@@ -1,6 +1,6 @@
 <?php
 
-namespace AppBundle\Entity;
+namespace AppBundle\Entity\Legacy;
 
 /**
  * Runde
