+++ src/AppBundle/Entity/Legacy/Tipp.php
@@ -1,6 +1,6 @@
 <?php
 
-namespace AppBundle\Entity;
+namespace AppBundle\Entity\Legacy;
 
 /**
  * Tipp
